fix(api,runner): change allocated metrics to double precision

+++ apps/runner/pkg/api/dto/info.go
@@ -8,9 +8,9 @@ type RunnerMetrics struct {
 	CurrentCpuUsagePercentage    float64 `json:"currentCpuUsagePercentage"`
 	CurrentMemoryUsagePercentage float64 `json:"currentMemoryUsagePercentage"`
 	CurrentDiskUsagePercentage   float64 `json:"currentDiskUsagePercentage"`
-	CurrentAllocatedCpu          int64   `json:"currentAllocatedCpu"`
-	CurrentAllocatedMemoryGiB    int64   `json:"currentAllocatedMemoryGiB"`
-	CurrentAllocatedDiskGiB      int64   `json:"currentAllocatedDiskGiB"`
+	CurrentAllocatedCpu          float64 `json:"currentAllocatedCpu"`
+	CurrentAllocatedMemoryGiB    float64 `json:"currentAllocatedMemoryGiB"`
+	CurrentAllocatedDiskGiB      float64 `json:"currentAllocatedDiskGiB"`
 	CurrentSnapshotCount         int     `json:"currentSnapshotCount"`
 	CurrentStartedSandboxes      int64   `json:"currentStartedSandboxes"`
 } //	@name	RunnerMetrics
